% Generated by roxygen2 (4.0.2): do not edit by hand
\name{neg_loglik_deriv}
\alias{neg_loglik_deriv}
\title{Calculate the gradient negative loglikelihood for logistic regression}
\usage{
neg_loglik_deriv(coefficients, response, design)
}
\arguments{
\item{coefficients}{a vector. The coefficients of the model, length must be
the same as the number of columns in the design matrix}

\item{response}{a vector. The resonse of the data, length must be the same as
the number of rows in the design matrix}

\item{design}{numeric matrix. The data used to fit the loglikelihood}
}
\description{
Calculate the gradient negative loglikelihood for logistic regression.
Used for optimization of the likelihood.
}
\author{
Janek Thomas, Philipp Roesch
}

